Tests/LibWeb/Text/input/css/nth-child-extreme-an-plus-b.html

<!DOCTYPE html>
<script src="../include.js"></script>
<ol>
    <li id="first"></li>
    <li id="second"></li>
    <li id="third"></li>
    <li id="fourth"></li>
    <li id="fifth"></li>
</ol>
<script>
    test(() => {
        const selectors = [
            ":nth-child(2n-2147483648)",
            ":nth-child(-2147483648n+1)",
            ":nth-child(2147483647n+1)",
            ":nth-child(0n+3)",
            ":nth-child(-2n+5)",
            ":nth-child(n-2)",
            ":nth-child(n-2147483648)",
            ":nth-last-child(n-2)",
            ":nth-of-type(n-2)",
        ];
        for (const id of ["first", "second", "third", "fourth", "fifth"]) {
            const element = document.getElementById(id);
            const results = selectors.map(selector => element.matches(selector));
            println(`${id}: ${results.join(",")}`);
        }
    });
</script>
