Tests/LibWeb/Text/expected/css/nth-child-extreme-an-plus-b.txt

first: false,true,true,false,true,true,true,true,true
second: true,false,false,false,false,true,true,true,true
third: false,false,false,true,true,true,true,true,true
fourth: true,false,false,false,false,true,true,true,true
fifth: false,false,false,false,true,true,true,true,true
